requirements.txt
torch==1.9.0
torchvision==0.10.0
Pillow==8.4.0
streamlit==0.87.0

